+++ angr_c/angr_c/memory_mixins/pages/decomposer.hpp
@@ -4,8 +4,8 @@
 #include <pybind11/pybind11.h>
 #include <stdint.h>
 #include <utility>
-#include "sim_memory_object.hpp"
-#include "endness.hpp"
+#include "../sim_memory_object.hpp"
+#include "../endness.hpp"
 
 namespace py = pybind11;
 
@@ -16,7 +16,7 @@ namespace angr_c
 	public:
 		Decomposer(uint64_t addr, py::object data, uint8_t byte_width, Endness endness, uint64_t page_addr); // TODO: label
 		virtual ~Decomposer();
-		virtual std::pair<SimMemoryObject*, uint64_t> yield(uint64_t size) = 0;
+		virtual std::pair<SimMemoryObject*, uint32_t> yield(uint32_t size) = 0;
 
 	protected:
 		uint64_t m_addr;
@@ -30,7 +30,7 @@ namespace angr_c
 	public:
 		MemoryObjectDecomposer(uint64_t addr, py::object data, uint8_t byte_width, Endness endness, uint64_t page_addr); // TODO: label
 		~MemoryObjectDecomposer();
-		std::pair<SimMemoryObject*, uint64_t> yield(uint64_t size);
+		std::pair<SimMemoryObject*, uint32_t> yield(uint32_t size);
 
 	private:
 		SimMemoryObject* m_memobj;
