+++ angr_c/angr_c/memory_mixins/sim_memory_object.cpp
@@ -3,7 +3,7 @@
 
 namespace angr_c
 {
-	SimMemoryObject::SimMemoryObject(py::object ast, uint64_t base, Endness endness, uint64_t byte_width, uint64_t length)
+	SimMemoryObject::SimMemoryObject(py::object ast, uint64_t base, Endness endness, uint8_t byte_width, uint32_t length)
 		: m_ast(ast), m_bytes(NULL), m_base(base), m_endness(endness), m_byte_width(byte_width)
 	{
 		if (length == 0) {
@@ -19,7 +19,7 @@ namespace angr_c
 		m_length = length;
 	}
 
-	SimMemoryObject::SimMemoryObject(uint8_t* bytes, uint64_t base, Endness endness, uint64_t byte_width, uint64_t length)
+	SimMemoryObject::SimMemoryObject(uint8_t* bytes, uint64_t base, Endness endness, uint8_t byte_width, uint32_t length)
 		: m_ast(py::none()), m_bytes(bytes), m_base(base), m_endness(endness), m_byte_width(byte_width), m_length(length)
 	{
